Mark Down update

+++ milestones/milestone4/cst339/src/main/java/com/gcu/services/InMemoryProductService.java
@@ -20,7 +20,7 @@ public class InMemoryProductService implements ProductService {
         Product p = new Product(
                 form.getName(),
                 form.getDescription(),
-                form.getPrice(),
+                form.getPrice().intValue(),
                 form.getQuantity()
         );
         products.add(p);
